Move data models to dedicated folder
