+++ GraduationTracker/GraduationTracker/Models/Standing.cs
@@ -1,0 +1,29 @@
+﻿namespace GraduationTracker.Models
+{
+    /// <summary>
+    /// Represents the academic standing of a student.
+    /// </summary>
+    public enum Standing
+    {
+        /// <summary>
+        /// No standing.
+        /// </summary>
+        None,
+        /// <summary>
+        /// With highest praise.
+        /// </summary>
+        SummaCumLaude,
+        /// <summary>
+        /// With great praise.
+        /// </summary>
+        MagnaCumLaude,
+        /// <summary>
+        /// Average standing.
+        /// </summary>
+        Average,
+        /// <summary>
+        /// Remedial standing.
+        /// </summary>
+        Remedial
+    }
+}
