--- GraduationTracker/GraduationTracker/Student.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GraduationTracker
-{
-    public class Student
-    {
-        public int Id { get; set; }
-        public Course[] Courses { get; set; }
-        public STANDING Standing { get; set; } = STANDING.None;
-    }
-}
